[feat/viewmode_state_handler]: PR 대응
