+++ app/src/main/java/com/teampophory/pophory/data/repository/photo/DefaultPhotoRepository.kt
@@ -1,22 +1,30 @@
 package com.teampophory.pophory.data.repository.photo
 
 import com.teampophory.pophory.common.image.ContentUriRequestBody
+import com.teampophory.pophory.common.okhttp.toPlainRequestBody
 import com.teampophory.pophory.data.model.photo.Studio
 import com.teampophory.pophory.data.network.model.album.PhotoListResponse
 import com.teampophory.pophory.data.network.service.PhotoService
-import com.teampophory.pophory.network.PhotoNetworkDataSource
+import okhttp3.RequestBody
+import retrofit2.HttpException
 import javax.inject.Inject
 
 class DefaultPhotoRepository @Inject constructor(
-    private val photoService: PhotoService,
-    private val photoNetworkDataSource: PhotoNetworkDataSource
+    private val photoService: PhotoService
 ) : PhotoRepository {
     override suspend fun getPhotos(id: Int): Result<PhotoListResponse> {
         return runCatching { photoService.getPhotos(id) }
     }
 
     override suspend fun deletePhoto(photoId: Long): Result<Unit> {
-        return runCatching { photoNetworkDataSource.deletePhoto(photoId) }
+        return runCatching {
+            val response = photoService.deletePhoto(photoId)
+            if (response.isSuccessful) {
+                response.body()
+            } else {
+                throw HttpException(response)
+            }
+        }
     }
 
     override suspend fun getStudios(): Result<List<Studio>> {
@@ -29,8 +37,19 @@ class DefaultPhotoRepository @Inject constructor(
         studioId: Long,
         photo: ContentUriRequestBody
     ): Result<Unit> {
+        val file = photo.toFormData("photo")
+        val request = HashMap<String, RequestBody>().apply {
+            put("albumId", albumId.toString().toPlainRequestBody())
+            put("takenAt", takenAt.toPlainRequestBody())
+            put("studioId", studioId.toString().toPlainRequestBody())
+        }
+        val response = photoService.addPhoto(file, request)
         return runCatching {
-            photoNetworkDataSource.addPhoto(albumId, takenAt, studioId, photo)
+            if (response.isSuccessful) {
+                response.body()
+            } else {
+                throw HttpException(response)
+            }
         }
     }
 }
